move router into index
