+++ index.spec.ts
@@ -7,9 +7,13 @@ import {
     RouteDefs,
 } from "mithril";
 
-import index from "./index";
+import * as m from "mithril/render/hyperscript";
 
-describe(index.name, () => {
+import middleware, {
+    router,
+} from "./index";
+
+describe(middleware.name, () => {
     const view = () => "";
     const defs = {
         "/": view,
@@ -37,14 +41,14 @@ describe(index.name, () => {
     }
     const MockNext = () => () => undefined;
     it("should return RequestHandler function", () => {
-        expect(index(defs)).to.be.a("function");
+        expect(middleware(defs)).to.be.a("function");
     });
     it("should returned RequestHandler require 3 arguments", () => {
-        expect(index(defs)).to.have.property("length").equal(3);
+        expect(middleware(defs)).to.have.property("length").equal(3);
     });
     it("should be working with express", () => {
         // @TODO add some spies
-        const handler = index(defs);
+        const handler = middleware(defs);
         for (const path of paths) {
             const req = new MockReq(path);
             const res = new MockRes();
@@ -53,7 +57,7 @@ describe(index.name, () => {
         }
     });
     it("should have working html function", () => {
-        const handler = index(defs, {
+        const handler = middleware(defs, {
             html: (partial) => Promise.resolve(partial).then((val) => {
                 expect(val).to.be.a("string");
                 return val;
@@ -65,5 +69,35 @@ describe(index.name, () => {
             const next = MockNext();
             handler(req as any, res as any, next as any);
         }
+    });
+});
+
+describe(router.name, () => {
+    const view = () => m("#", "test");
+    const defs = {
+        "/": {
+            render: view,
+        },
+        "/abc/test": view,
+        "/abc/:test": view,
+        "/abc/:test...": view,
+    } as RouteDefs;
+    it("should return Promise", () => {
+        return expect(router(defs, "/")).to.be.instanceOf(Promise);
+    });
+    it("should resolve to string", async () => {
+        expect(await router(defs, "/abc/test")).to.be.a("string");
+    });
+    it("should throw exception if route does not exists", async () => {
+        let err;
+        try {
+            await router(defs, "not-exists");
+        } catch (e) {
+            err = e;
+        }
+        expect(err).to.be.instanceOf(Error);
+    });
+    it("should have working defaut route", async () => {
+        expect(await router(defs, "not-exists", "/abc/abc")).to.be.a("string");
     });
 });
