+++ index.ts
@@ -1,17 +1,51 @@
 import {
     RequestHandler,
 } from "express";
+
 import {
     RouteDefs,
 } from "mithril";
 
-import router from "./router";
+import {
+    parsePath,
+} from "./parse";
 
-export type HtmlFunction = (partial: string | Promise<string>) => string | Promise<string>;
+import render from "./render";
+
+export async function router(
+    routes: RouteDefs,
+    path: string,
+    defaultPath: string = path,
+): Promise<string> {
+    const params: { [_: string]: any } = {};
+    const pathname = parsePath(path, params, params);
+    // tslint:disable-next-line:forin
+    for (const route in routes) {
+        const matcher = new RegExp(
+            `^${route.replace(/:[^\/]+?\.{3}/g, "(.*?)").replace(/:[^\/]+/g, "([^\\/]+)")}\/?$`,
+        );
+        const matchArr = pathname.match(matcher);
+        if (matchArr !== null) {
+            const keys = route.match(/:[^\/]+/g) || [];
+            const [match, ...values] = matchArr;
+            // tslint:disable-next-line:forin
+            for (const i in keys) {
+                const key = keys[i];
+                const value = values[i];
+                params[key.replace(/:|\./g, "")] = decodeURIComponent(value);
+            }
+            return await render(routes[route], params, path);
+        }
+    }
+    if (path !== defaultPath) {
+        return await router(routes, defaultPath);
+    }
+    throw new Error(`Could not resolve route '${path}'`);
+}
 
 export interface IOptions {
-    html?: HtmlFunction;
-    defaultRoute?: string;
+    html: (partial: string | Promise<string>) => string | Promise<string>;
+    defaultRoute: string;
 }
 
 export function mithrilExpressMiddleware(
@@ -19,7 +53,7 @@ export function mithrilExpressMiddleware(
     {
         html = (partial) => partial,
         defaultRoute,
-    }: IOptions = {},
+    }: Partial<IOptions> = {},
 ): RequestHandler {
     return async (req, res, next) => {
         try {
